fix naimg and imports
