+++ packages/types/surveys/types.ts
@@ -2477,14 +2477,3 @@ export const ZSurveyRecallItem = z.object({
 });
 
 export type TSurveyRecallItem = z.infer<typeof ZSurveyRecallItem>;
-
-export const ZSurveyCopyFormValidation = z.object({
-  products: z.array(
-    z.object({
-      product: z.string(),
-      environments: z.array(z.string()),
-    })
-  ),
-});
-
-export type TSurveyCopyFormData = z.infer<typeof ZSurveyCopyFormValidation>;
